Rename the package to logger
