+++ logger/logger_test.go
@@ -1,4 +1,4 @@
-package log
+package logger
 
 import (
 	"bytes"
